refactor: cleanup token generation logic
